+++ app/src/main/java/com/example/playlistmaker/search/ui/TrackSearchViewModel.kt
@@ -9,6 +9,7 @@ import com.example.playlistmaker.helpers.AppConstants
 import com.example.playlistmaker.search.data.TrackDto
 import com.example.playlistmaker.search.domain.TrackInteractor
 import com.example.playlistmaker.util.debounce
+import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
 
 class TrackSearchViewModel(private val trackInteractor: TrackInteractor) : ViewModel() {
@@ -17,11 +18,10 @@ class TrackSearchViewModel(private val trackInteractor: TrackInteractor) : ViewM
     val searchScreenState: LiveData<SearchScreenState> = screenState
     private val showToast = SingleLiveEvent<String>()
     private var isClickable = true
+    private var searchJob: Job? = null
 
     var query = mutableStateOf("")
         private set
-
-    fun observeState(): LiveData<SearchScreenState> = screenState
 
     fun observeShowToast(): LiveData<String> = showToast
 
@@ -38,6 +38,7 @@ class TrackSearchViewModel(private val trackInteractor: TrackInteractor) : ViewM
             viewModelScope,
             true
         ) { changedText ->
+            searchJob?.cancel()
             getTracks(changedText)
         }
 
@@ -62,9 +63,9 @@ class TrackSearchViewModel(private val trackInteractor: TrackInteractor) : ViewM
     }
 
     fun getTracks(query: String) {
-        if (query.isNotEmpty()) {
+        if (query.isNotEmpty() && query == this.query.value) {
             renderState(SearchScreenState.Loading)
-            viewModelScope.launch {
+            searchJob = viewModelScope.launch {
                 trackInteractor.searchSongs(query).collect { pair ->
                     processResult(pair.first, pair.second)
                 }
@@ -116,8 +117,9 @@ class TrackSearchViewModel(private val trackInteractor: TrackInteractor) : ViewM
     }
 
     fun clearSearch() {
-        val historyTracks = showHistory()
         query.value = ""
+        val historyTracks = showHistory()
+        searchJob?.cancel()
         if (historyTracks.isNotEmpty()) {
             renderState(SearchScreenState.ShowHistory(historyTracks))
         } else {
